+++ sql/update/20200712.sql
@@ -1,1 +1,2 @@
 ALTER TABLE t_remind ADD COLUMN task_id INTEGER;
+ALTER TABLE task_queue ADD COLUMN remind_id INTEGER;